Add OPAMP modeling through equivalent circuit
This commit introduces the OPMODEL circuit component to simulate an
OPAMP. No actual new component types are added, OPMODEL is instead a
'macro' that will generate an equivalent circuit.
A feeback resistor is hardcoded between the inverting terminal of the
OPAMP and its output terminal, the resistance value is specified in the
Value field of the OPMODEL itself.

The necessary elements of the equivalent circuit are appended to the
current netlist and then parsed as usual.